+++ Source/filter.py
@@ -10,33 +10,30 @@ import xbmc
 import json
 import searchOptions
 
-# {"field": "genre", "operator": "contains", "value": ["", "", ""]}
-# {"field": "genre", "operator": "is", "value": ""}
-def _genre_filter_(genreList):
+# {"field": name, "operator": "contains", "value": ["", "", ""]}
+# {"field": name, "operator": "is", "value": ""}
+def _filterList_(name,valueList):
     defaultReturn = {}
-    if len(genreList) == 1:
-        defaultReturn = {"field": "genre", "operator": "is", "value": genreList[0]}
-    elif len(genreList) > 1:
-        for x in genreList:
+    if len(valueList) == 1:
+        defaultReturn = {"operator": "is", "field": name, "value": valueList[0]}
+    elif len(valueList) > 1:
+        for x in valueList:
             if defaultReturn == {}:
-                defaultReturn = x
-            else:
-                defaultReturn 
+                defaultReturn = []
+            defaultReturn.append(x)
+        defaultReturn = {"operator": "contains", "field": name, "value": defaultReturn}
     return defaultReturn
 
 # Filter function
 ## Pass in the Search Options
 def filter(options: searchOptions):
-    genreFilter = _genre_filter_(options.getGenre())
-    tagFilter = _tag_filter_(options.getTag())
-    castFilter = _cast_filter_(options.getCast())
-    directorFilter = _director_filter_(options.getDirector())
-    includeFilter = _include_filter_(options.getInclude())
-    excludeFilter = _exclude_filter_(options.getExclude())
-    lengthFilter = _length_filter_(options.getLength())
-    yearFilter = _year_filter_(options.getYear())
-    studioFilter = _studio_filter_(options.getStudio())
-    ratingFilter = _rating_filter_(options.getRating())
+    genreFilter = _filterList_('genre', options.getGenre())
+    tagFilter = _filterList_('tag', options.getTag())
+    castFilter = _filterList_('actor', options.getCast())
+    directorFilter = _filterList_('director', options.getDirector())
+    yearFilter = _filterList_('year', options.getYear())
+    studioFilter = _filterList_('studio', options.getStudio())
+    ratingFilter = _filterList_('mpaarating', options.getRating())
 
     filterList = []
     if genreFilter != {}:
@@ -47,18 +44,14 @@ def filter(options: searchOptions):
         filterList.append(castFilter)
     if directorFilter != {}:
         filterList.append(directorFilter)
-    if includeFilter != {}:
-        filterList.append(includeFilter)
-    if excludeFilter != {}:
-        filterList.append(excludeFilter)
-    if lengthFilter != {}:
-        filterList.append(lengthFilter)
-    if yearFilter != {}:
-        filterList.append(yearFilter)
+    '''if yearFilter != {}:
+        filterList.append(yearFilter)'''
     if studioFilter != {}:
         filterList.append(studioFilter)
     if ratingFilter != {}:
         filterList.append(ratingFilter)
+
+    print (filterList)
 
     # if movie type
     # function call to get movie list
@@ -66,31 +59,57 @@ def filter(options: searchOptions):
     # function call to get episode list
     # combine show lists
     # trigger playback functions
-    videoList = []
+    mediaTypes = options.getMediaType()
+    movieList = []
+    episodeList = []
+    for x in mediaTypes:
+        if(x == 'movie'):
+            command = {
+                "jsonrpc": "2.0",
+                "method": "VideoLibrary.GetMovies",
+                "params": {
+                    "filter": { "and" : filterList }, # get_genre_filer() = {"field": "genre", "operator": "is", "value": genre}
+                    "properties": ["art", "thumbnail", "playcount", "file", "runtime", "genre", "tag", "cast", "director", "year", "studio", "mpaa", "rating", "userrating"],
+                    "sort": {"order": "ascending", "method": "label"}
+                },
+                "id": "library"}
+            print(command)
+            movieList = json.loads(xbmc.executeJSONRPC(json.dumps(command)))
+        elif(x == 'episode'):
+            command = {
+                "jsonrpc": "2.0",
+                "method": "VideoLibrary.GetEpisodes",
+                "params": {
+                    "filter": { "or" : filterList }, # get_genre_filer() = {"field": "genre", "operator": "is", "value": genre}
+                    "properties": ["art", "rating", "thumbnail", "playcount", "file", "genre"],
+                    "sort": {"order": "ascending", "method": "label"}
+                },
+                "id": "library"}
+            print(command)
+            episodeList = json.loads(xbmc.executeJSONRPC(json.dumps(command)))
+
+    #videoList = episodeList + movieList
+    print(episodeList)
+    print(movieList)
+
+    '''videoList = filterIncExc(videoList, options.getInclude(), options.getExclude())
+    videoList = filterLength(videoList, options.getInclude(), options.getExclude())
+
     if options.getPBFunction() == 1:
         playOne(videoList)
     elif options.getPBFunction() == 2:
         showList(videoList)
     elif options.getPBFunction() == 3:
-        loopPlay(videoList)
+        loopPlay(videoList)'''
 
-    command = {
-        "jsonrpc": "2.0",
-        "method": "VideoLibrary.GetEpisodes",
-        "params": {
-            "filter": { "or" : filterList }, # get_genre_filer() = {"field": "genre", "operator": "is", "value": genre}
-            "properties": ["art", "rating", "thumbnail", "playcount", "file"],
-            "sort": {"order": "ascending", "method": "label"}
-        },
-        "id": "library"}
-    
-    response = xbmc.executeJSONRPC(json.dumps(command))
-    response_obj = json.loads(response)
-    if "result" in response_obj:
-        items = response_obj["result"]["episodes"]
-        items.extend(response_obj["result"]["movies"])
-        return items
-    else:
-        print("no item found")
-        return {}
 
+def unitTest(options: searchOptions):
+    print(_filterList_('genre', options.getGenre()))
+    print(_filterList_('tag', options.getTag()))
+    print(_filterList_('cast', options.getCast()))
+    print(_filterList_('director', options.getDirector()))
+    print(_filterList_('year', options.getYear()))
+    print(_filterList_('studio', options.getStudio()))
+    print(_filterList_('rating', options.getRating()))
+
+    filter(options)